reorganize files in collectors package
